[TASK] Finish setup splash  and home activity and redirect to auth from home. Also added icons and fonts.
