+++ app/src/main/java/com/opet/hasgarddetouro_app/SplashActivity.java
@@ -5,7 +5,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.util.Log;
+
+import com.opet.hasgarddetouro_app.activities.AuthenticationActivity;
 
 public class SplashActivity extends AppCompatActivity {
