+++ app/src/main/res/mipmap-anydpi-v26/ic_plus_flat_icon_round.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
+    <background android:drawable="@drawable/ic_plus_flat_icon_background"/>
+    <foreground android:drawable="@mipmap/ic_arrow_flat_icon_foreground"/>
+</adaptive-icon>
